fix: fix group type filter, refactor model extraction, move catalogs
