+++ src/app/features/private/community/components/community-page/community-page.component.css
@@ -8,3 +8,21 @@
     justify-content: center;
     padding: var(--spacing-medium) 0;
 }
+
+@media (max-width: 768px) {
+    .search-container {
+        flex: 1;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+    }
+
+    .paging-div {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        padding: var(--spacing-medium) 0;
+    }
+
+}
