--- src/app/features/private/community/services/groups-api/group-api.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { GroupApiService } from './group-api.service';
-
-describe('GroupApiService', () => {
-  let service: GroupApiService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(GroupApiService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
